Join neighbouring loops
Note that you need to be careful not to merge all the loops here - the deltas should be fully computed before applying them to the boids. If the last "apply updates" loop was merged with the "compute boid velocity updates" loop it would change the results (and the regression test would fail).